requirements.txt
Flask==2.0.2
Flask_SQLAlchemy==2.5.1
python-dotenv==0.19.2
SQLAlchemy==1.4.31
